// src/enums/finance.enum.ts
enum TYPE  {
    USER_PAYMENT = 'USER_PAYMENT',
    VEHICLE_PAYMENT = 'VEHICLE_PAYMENT',
    VENDOR_PAYMENT = 'VENDOR_PAYMENT',
    BANK_PAYMENT = 'BANK_PAYMENT',
    PAID_COMMISSION = 'PAID_COMMISSION',
    PENDING_COMMISSION = 'PENDING_COMMISSION',
    LAB_PAYMENT = 'LAB_PAYMENT',
    PAY_TO_ADMIN = 'PAY_TO_ADMIN',
    BANK_DEPOSIT = 'BANK_DEPOSIT',
    HOME_SAMPLE_REVENUE = 'HOME_SAMPLE_REVENUE',
    DOCTOR_PAY = 'DOCTOR_PAY',
    CUSTOMER_PAY = 'CUSTOMER_PAY',

}
enum PROCEDURE_OR_CONSULTANT { 
    PROCEDURE_FEE = 'PROCEDURE_FEE',
    CONSULTANT_FEE = 'CONSULTANT_FEE'
}

enum STATUS  {
    PENDING = 'PENDING',
    CANCELLED = 'CANCELLED',
    COMPLETED = 'COMPLETED',
}

enum PAYMENT_TYPE  {
    CASH = "CASH", // Rider
    CARD = "CARD", // Vendor
    CREDIT_CARD = "CREDIT_CARD", // Bank
    ONLINE_TRANSFER = "ONLINE_TRANSFER", // Bank
}

export  {
    TYPE,
    STATUS,
    PAYMENT_TYPE,
    PROCEDURE_OR_CONSULTANT
}
export default {
    TYPE,
    STATUS,
    PAYMENT_TYPE,
    PROCEDURE_OR_CONSULTANT
}